feat: Introduce FSTs
As per suggestion of BurntSushi in:

https://users.rust-lang.org/t/fast-string-lookup-in-a-single-str-containing-millions-of-unevenly-sized-substrings/98040

Don't need an FST (== map), only an FSA though (== set).

FST build times are fast, lookups are *okay* (actually slower than b4s
by c. a factor of 10), but FSTs also have compression built-in, which is
amazing as it's been the next problem on the list. Taking care of
compression equals out the relatively slow lookup times.