feat!: `grep`-like, recursive search mode
Breaking change in the sense that 'no actions' was
previously no-op, now no longer is.

There were also breaking changes to the library,
e.g. in the `Scoper` trait.

The 'search mode' is entered if no actions are
specified. No actions was previously just an
error, with input returned unchanged. It now
instead starts *searching* either stdin, if
provided, or files corresponding to the language
scoper provided, or, if given, files as per the
`--files` glob (which takes precedence over
language scoping heuristics). Search mode *also
requires a language to be specified*. If it
didn't, it'd just be a very crappy ripgrep. With
language grammar info, it becomes its own tool.

Lots of tests were written, which required a bunch
of fixes and extensions to the existing test
harness.

The CLI was also extended with a host of new
options.

The search mode itself was simple: essentially
just creating the `Style` `Action`, and giving the
`lines` method to `ScopedView`. Then print the
result out neatly.