build(tree-sitter-hcl): Remove `git` dependency
Publishing to crates.io requires all dependencies
be available on crates.io. The upstream crate for
`tree-sitter-hcl` is not. However, we can include
it as a submodule and massage the output(s) of
`tree-sitter generate` (specifically, the
`build.rs` and `lib.rs` it generates) to provide
just enough code in this repo to be able to
leverage the existing `tree-sitter-hcl`.

This allows us to remove my own git fork
dependency. I had forked the upstream repo to
update its `tree-sitter` dependency. That part
isn't needed anymore! However, it means we still
will only be able to use upstream
`tree-sitter-hcl` once it also updated its
`tree-sitter` dependency.