fix(language-scoping): Construct `TSQuery` only once
A query is static across the entire lifetime of a
run. It does not change, but it was `clone()`d on
each call to `scope_via_query`, which is
expensive. This change constructs the *actually*
required `TSQuery` only once (alongside its
"negative" counterpart, which also allows later
parts to just take a `&TSQuery` instead of `&mut
TSQuery`).

Closes #76