Flatten `mod scoping` into root
`mod scoping` is adding an unnecessary level to the mod hierarchy.
It is true that the sub-modules are involved in 'scoping',
but does that need to be taxonimized into the mod hierarchy?
I argue that flattening the `mod scoping` into the parent mod
reduces the complexity of the hierarchy without lossing meaning.