feat(cli): Dry run mode ( `--dry-run` flag) (#156)
This change adds a new flag `--dry-run` that when
given, prevents otherwise destructive file
overwrite operations and instead prints a rich
diff. The diff contains names of files which would
be modified, and a git-like (word) diff of what
modifications *would* be made.

The diff is human-readable, and not designed to be
easily machine-readable. It leverages the existing
search mode.

Closes #152.