chore(cli): Base-line work for `--[lang]-query-file` to read tree-sitter (scm) query files (#144)
This is a `feat` in terms of conventional commits, but committing as `chore` as this 
current implementation is not the final version yet. Things will change in breaking ways.

Once the final changes are in place (dedicated argument of (something like) `--<lang>-query-file`, leaving
the existing `--<lang>-query` as it worked before this change), that will be committed
as a `feat` and contribute to release notes etc. Nothing in this commit is supposed
to be user-facing, and not intended to make it into a release (next one is 0.13.4 at the
time of writing).