feat(cli): Parseable output format (with auto-detection)
This change adds a machine-readable output format.
`srgn` tries to detect automatically if it should
be used by probing if standard output is a TTY. If
it is not, this new format is used. If it is, the
known, current format is used (which remains
unchanged).

The output format is designed for further
processing in downstream scripts. It works
similarly to ripgrep.

This is all a bit ad-hoc and crappy - something
structured and well-known like JSON might be nice.
It could be added, though the current approach is
already pretty spaghetti.