feat(cli)!: Replacement action argument forced last
This change introduces a tiny change: the second
positional argument to `srgn` is the
*replacement*. This was somewhat dangerous, as
something like `srgn A B` would be destructive,
even if the user meant `srgn 'A B'`.

The second argument is now **forced to come in
last position, disambiguated by `--`**. This
forces usage like `srgn A -- B`, making
replacement more explicit and thus safer, as it's
harder to trigger accidentally.

An invocation like `srgn A B` now fails. Spaces
need to be quoted: `srgn 'A B'` works.

This is a **breaking change** for the CLI, though
the adjustment is very small.

Closes #149 .